% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbprofile.R
\name{confint.tmbprofile}
\alias{confint.tmbprofile}
\title{Profile based confidence intervals.}
\usage{
\method{confint}{tmbprofile}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{Output from \code{\link{tmbprofile}}.}

\item{parm}{Not used}

\item{level}{Confidence level.}

\item{...}{Not used}
}
\value{
Lower and upper limit as a matrix.
}
\description{
Calculate confidence interval from a likelihood profile.
}
