% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB.R
\name{dynlib}
\alias{dynlib}
\title{Add dynlib extension}
\usage{
dynlib(name)
}
\arguments{
\item{name}{Library name without extension}
}
\value{
Character
}
\description{
Add the platform dependent dynlib extension. In order for examples
to work across platforms DLLs should be loaded by
\code{dyn.load(dynlib("name"))}.
}
