% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB.R
\name{GK}
\alias{GK}
\title{Gauss Kronrod configuration}
\usage{
GK(...)
}
\arguments{
\item{...}{See source code}
}
\description{
Helper function to specify parameters used by the Gauss Kronrod
integration available through the argument \code{integrate} to
\code{MakeADFun}.
}
