% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc.hmc}
\alias{mcmc.hmc}
\title{[BETA VERSION] Draw MCMC samples from a model posterior using a Hamiltonian sampler.}
\usage{
mcmc.hmc(nsim, L, eps, fn, gr, params.init, covar = NULL,
  diagnostic = FALSE)
}
\arguments{
\item{nsim}{The number of samples to return.}

\item{L}{The number of leapfrog steps to take. The NUTS algorithm does
not require this as an input. If L is 1 this function will perform
Langevin sampling.}

\item{eps}{The length of the leapfrog steps.}

\item{fn}{A function that returns the log of the posterior density. This
function should not return the negative log, following the notation of
Hoffman and Gelman (2014).}

\item{gr}{A function that returns a vector of gradients of the log of
the posterior density (same as with \code{fn}).}

\item{params.init}{A vector of initial parameter values.}

\item{diagnostic}{Whether to return a list of diagnostic metrics about
the chain. Useful for assessing efficiency and tuning chain.}
}
\value{
If \code{diagnostic} is FALSE (default), returns a matrix of
\code{nsim} samples from the posterior. Otherwise returns a list
containing samples ('par'), proposed samples ('par.proposed'), vector of
which were accepted ('accepted'), and the total function and gradient
calls ('n.calls'), which for this algorithm is \code{nsim}*(\code{L}+2)
}
\description{
[BETA VERSION] Draw MCMC samples from a model posterior using a Hamiltonian sampler.
}
\references{
Neal, R. M. 2011. MCMC using Hamiltonian dynamics.
}

