% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc.nuts}
\alias{mcmc.nuts}
\title{[BETA VERSION] Draw MCMC samples from a model posterior using the
No-U-Turn (NUTS) sampler from Hoffman and Gelman (2014).}
\usage{
mcmc.nuts(nsim, fn, gr, params.init, eps, diagnostic = FALSE)
}
\arguments{
\item{nsim}{The number of samples to return.}

\item{fn}{A function that returns the log of the posterior density. This
function should not return the negative log, following the notation of
Hoffman and Gelman (2014).}

\item{gr}{A function that returns a vector of gradients of the log of
the posterior density (same as with \code{fn}).}

\item{params.init}{A vector of initial parameter values.}

\item{eps}{The length of the leapfrog steps.}

\item{diagnostic}{Whether to return a list of diagnostic metrics about
the chain. Useful for assessing efficiency and tuning chain.}
}
\value{
If \code{diagnostic} is FALSE (default), returns a matrix of
\code{nsim} samples from the posterior. Otherwise returns a list
containing samples ('par'),  vector of steps taken at each iteration
('steps.taken'), and the total function and gradient
calls ('n.calls').
}
\description{
[BETA VERSION] Draw MCMC samples from a model posterior using the
No-U-Turn (NUTS) sampler from Hoffman and Gelman (2014).
}
\details{
This is the 'efficient' NUTS algorithm but does not use dual
averaging to tune \code{eps}, so it must be specified.
}
\references{
Hoffman and Gelman (2014). The No-U-Turn sampler: Adaptively
setting path lengths in Hamiltonian Monte Carlo.
}

