% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc.rwm}
\alias{mcmc.rwm}
\title{[BETA VERSION] Draw MCMC samples from a model posterior using a
Random Walk Metropolis (RWM) sampler.}
\usage{
mcmc.rwm(nsim, fn, params.init, alpha = 1, covar = NULL,
  diagnostic = FALSE)
}
\arguments{
\item{nsim}{The number of samples to return.}

\item{fn}{A function that returns the log of the posterior density.}

\item{params.init}{A vector of initial parameter values.}

\item{alpha}{The amount to scale the proposal, i.e,
Xnew=Xcur+alpha*Xproposed where Xproposed is generated from a mean-zero
multivariate normal. Varying \code{alpha} varies the acceptance rate.}

\item{covar}{A covariance matrix to be used in generating multivariate
normal proposals. A value of NULL (default) indicates to use iid
standard normal.}

\item{diagnostic}{Whether to return a list of diagnostic metrics about
the chain. Useful for assessing efficiency and tuning chain.}
}
\value{
If \code{diagnostic} is FALSE (default), returns a matrix of
\code{nsim} samples from the posterior. Otherwise returns a list
containing samples ('par'), proposed samples ('par.proposed'), vector of
which were accepted ('accepted'), and the total function calls
('n.calls'), which for this algorithm is \code{nsim}
}
\description{
[BETA VERSION] Draw MCMC samples from a model posterior using a
Random Walk Metropolis (RWM) sampler.
}
\details{
This
}

