% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metro.stations.R
\name{metro.stations}
\alias{metro.stations}
\title{Get data from the metro service in Barcelona}
\usage{
# Without the line argument, the function retrieves
# data from all metro stations
metro.stations(app_id, app_key)

# Metro stations corresponding to the specified line
metro.stations(app_id, app_key,
               line = c(1, 2, 3, 4, 5, 11, 91,
                        94, 99, 101, 104))
}
\arguments{
\item{app_id}{character}

\item{app_key}{character}

\item{line}{numeric}
}
\description{
The \code{metro.stations()} function opens a connection
to the \href{https://developer.tmb.cat/}{TMB API} to retrieve data related
to the metro stations and lines in Barcelona.
}
\details{
\itemize{
  \item{To execute the function you need the credentials \code{app_id} and
  \code{app_key}.
   Register on the \href{https://developer.tmb.cat/docs/getting-started}{TMB API portal}
   and create an application to obtain your credentials}
 \item{If you want to retrieve all the metro stations
 and their resources, execute the function without arguments.}
 \item{You can filter by metro line and obtain a subset.}
 }
}
\author{
Xavier Vivancos
}
