% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB.R
\name{newtonOption}
\alias{newtonOption}
\title{Set newton options for a model object.}
\usage{
newtonOption(obj, ...)
}
\arguments{
\item{obj}{Object from \code{\link{MakeADFun}} for which to change settings.}

\item{...}{Parameters for the \code{\link{newton}} optimizer to set.}
}
\value{
List of updated parameters.
}
\description{
Inner-problem options can be set for a model object using this
function.
}
