% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB.R
\name{openmp}
\alias{openmp}
\title{Control number of openmp threads.}
\usage{
openmp(n = NULL)
}
\arguments{
\item{n}{Requested number of threads, or \code{NULL} to just read the current value.}
}
\value{
Number of threads.
}
\description{
Control number of openmp threads.
}
