% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{print.checkConsistency}
\alias{print.checkConsistency}
\title{Print output from \code{\link{checkConsistency}}}
\usage{
\method{print}{checkConsistency}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{checkConsistency}}}

\item{...}{Not used}
}
\description{
Print diagnostics output from \code{\link{checkConsistency}}
}
