% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdreport.R
\name{print.sdreport}
\alias{print.sdreport}
\title{Print brief model summary}
\usage{
\method{print}{sdreport}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{sdreport}}}

\item{...}{Not used}
}
\description{
Print parameter estimates and give convergence diagnostic based on
gradient and Hessian.
}
