% Generated by roxygen2 (4.0.1): do not edit by hand
\name{runExample}
\alias{runExample}
\title{Run one of the test examples.}
\usage{
runExample(name = NULL, all = FALSE, thisR = TRUE, clean = FALSE,
  exfolder = NULL, ...)
}
\arguments{
\item{name}{Character name of example.}

\item{all}{Run all the test examples?}

\item{thisR}{Run inside this R?}

\item{clean}{Cleanup before compile?}

\item{exfolder}{Alternative folder with examples.}

\item{...}{Passed to \code{compile}.}
}
\description{
Compile and run a test example (\code{runExample()} shows all available examples).
}

