% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdreport.R
\name{summary.sdreport}
\alias{summary.sdreport}
\title{summary tables of model parameters}
\usage{
\method{summary}{sdreport}(object, select = c("all", "fixed", "random",
  "report"), p.value = FALSE, ...)
}
\arguments{
\item{object}{Output from \code{\link{sdreport}}}

\item{select}{Parameter classes to select. Can be any subset of
\code{"fixed"} (\eqn{\hat\theta}), \code{"random"} (\eqn{\hat u}) or
\code{"report"} (\eqn{\phi(\hat u,\hat\theta)}) using notation as
\code{\link{sdreport}}.}

\item{p.value}{Add column with approximate p-values}

\item{...}{Not used}
}
\value{
matrix
}
\description{
Extract parameters, random effects and reported variables along
with uncertainties and optionally Chi-square statistics. Bias
corrected quantities are added as additional columns if available.
}
