% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_pre.R
\name{tmb_pre}
\alias{tmb_pre}
\title{TMBpredict}
\usage{
tmb_pre(ttype, mut, panel.bed, ftype)
}
\arguments{
\item{ttype}{Tumor type}

\item{mut}{Detected mutations in VCF format}

\item{panel.bed}{Panel region file in BED format}

\item{ftype}{The type of input mutation file,"s" for single VCF file,"m" for multiple VCF files compressed with tar.gz.}
}
\value{
The adjusted TMB value and correlation figure
}
\description{
Correct panel TMB evaluation by modeling adjustment.
}
\examples{
tmb_pre("COAD","sample.vcf","panel.bed")
}
