% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBpredict.R
\name{TMBpredict}
\alias{TMBpredict}
\title{TMBpredict}
\usage{
TMBpredict(ttype, mut, panel.bed, ftype)
}
\arguments{
\item{ttype}{Tumor type}

\item{mut}{Detected mutations in VCF format}

\item{panel.bed}{Panel region file in BED format}

\item{ftype}{The type of input mutation file,"s" for single VCF file,"m" for multiple VCF files compressed with tar.gz.}
}
\value{
The adjusted TMB value and correlation figure
}
\description{
Correct panel TMB evaluation by modeling adjustment.
}
\examples{
TMBpredict("COAD","sample.vcf","panel.bed","s")
}
