% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc.hmc}
\alias{mcmc.hmc}
\title{[BETA VERSION] Draw MCMC samples from a model posterior using a
Hamiltonian sampler.}
\usage{
mcmc.hmc(nsim, fn, gr, params.init, L, eps = NULL, covar = NULL,
  delta = 0.5, Madapt = NULL, diagnostic = FALSE)
}
\arguments{
\item{nsim}{The number of samples to return.}

\item{fn}{A function that returns the log of the posterior density.}

\item{gr}{A function that returns a vector of gradients of the log of
the posterior density (same as \code{fn}).}

\item{params.init}{A vector of initial parameter values.}

\item{L}{The number of leapfrog steps to take. The NUTS algorithm does
not require this as an input. If \code{L=1} this function will perform
Langevin sampling. In some contexts \code{L} can roughly be thought of
as a thinning rate.}

\item{eps}{The length of the leapfrog steps. If a numeric value is
passed, it will be used throughout the entire chain. A \code{NULL}
value will initiate adaptation of \code{eps} using the dual averaging
algorithm during the first \code{Madapt} steps.}

\item{covar}{An optional covariance matrix which can be used to improve
the efficiency of sampling. The lower Cholesky decomposition of this
matrix is used to transform the parameter space. If the posterior is
approximately multivariate normal and \code{covar} approximates the
covariance, then the transformed parameter space will be close to
multivariate standard normal. In this case the algorithm will be more
efficient, but there will be overhead in the matrix calculations which
need to be done at each step. The default of NULL specifies to not do
this transformation.}

\item{delta}{The target acceptance rate if using apative
\code{eps}. Defaults to 50\%.}

\item{Madapt}{An optional argument for how many iterations to adapt
\code{eps} in the dual averaging algorithm. A value of \code{NULL}
results in a default of \code{Madapt=nsim/2}.}

\item{diagnostic}{Whether to return a list of diagnostic metrics about
the chain. Useful for assessing efficiency and tuning chain.}
}
\value{
If \code{diagnostic} is FALSE (default), returns a matrix of
\code{nsim} samples from the posterior. Otherwise returns a list
containing samples ('par'), proposed samples ('par.proposed'), vector of
which were accepted ('accepted'), and the total function and gradient
calls ('n.calls'), which for this algorithm is \code{nsim*(L+2)}.
}
\description{
[BETA VERSION] Draw MCMC samples from a model posterior using a
Hamiltonian sampler.
}
\details{
This function implements algorithm 5 of Hoffman and Gelman
(2014), which includes adaptive step sizes (\code{eps}) via an algorithm
called dual averaging. In theory neither the step length nor step size
needs to be input by the user to obtain efficient sampling from the
posterior.
}
\references{
\itemize{
\item{Neal, R. M. (2011). MCMC using Hamiltonian dynamics. Handbook
of Markov Chain Monte Carlo.}
\item{Hoffman and Gelman (2014). The No-U-Turn sampler: Adaptively
setting path lengths in Hamiltonian Monte Carlo. J. Mach. Learn. Res.
15:1593-1623.}
}
}
\seealso{
\code{\link{mcmc}}, \code{\link{mcmc.nuts}}, \code{\link{mcmc.rwm}}
}

