% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB.R
\name{Rinterface}
\alias{Rinterface}
\title{Create minimal R-code corresponding to a cpp template.}
\usage{
Rinterface(file)
}
\arguments{
\item{file}{cpp template file.}
}
\description{
Create a skeleton of required R-code once the cpp template is ready.
}
\examples{
file <- system.file("examples/simple.cpp", package = "TMB")
Rinterface(file)
}
