% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/examples.R
\name{runExample}
\alias{runExample}
\title{Run one of the test examples.}
\usage{
runExample(name = NULL, all = FALSE, thisR = TRUE, clean = FALSE,
  exfolder = NULL, dontrun = FALSE, ...)
}
\arguments{
\item{name}{Character name of example.}

\item{all}{Run all the test examples?}

\item{thisR}{Run inside this R?}

\item{clean}{Cleanup before compile?}

\item{exfolder}{Alternative folder with examples.}

\item{dontrun}{Build only (don't run) and remove temporary object files ?}

\item{...}{Passed to \code{\link{compile}}.}
}
\description{
Compile and run a test example (\code{runExample()} shows all available examples).
}

