% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB.R
\name{runSymbolicAnalysis}
\alias{runSymbolicAnalysis}
\title{Run symbolic analysis on sparse Hessian}
\usage{
runSymbolicAnalysis(obj)
}
\arguments{
\item{obj}{Output from \code{MakeADFun}}
}
\description{
Aggressively tries to reduce fill-in of sparse Cholesky factor by
running a full suite of ordering algorithms. NOTE: requires a
specialized installation of the package. More information is
available at the package URL.
}
