% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{summary.checkConsistency}
\alias{summary.checkConsistency}
\title{Summarize output from \code{\link{checkConsistency}}}
\usage{
\method{summary}{checkConsistency}(object, na.rm = FALSE, ...)
}
\arguments{
\item{object}{Output from \code{\link{checkConsistency}}}

\item{na.rm}{Logical; Remove failed simulations ?}

\item{...}{Not used}
}
\value{
List of diagnostics
}
\description{
Summarize output from \code{\link{checkConsistency}}
}
