% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBpredict.R
\name{TMBpredict}
\alias{TMBpredict}
\title{TMBpredict}
\usage{
TMBpredict(ttype, mut, panel.bed, convert.type)
}
\arguments{
\item{ttype}{Tumor type}

\item{mut}{Detected mutations in VCF format. Also accept multiple VCF compressed in .tar.gz.}

\item{panel.bed}{Panel region file in BED format}

\item{convert.type}{Convert Panel to WES or reversely}
}
\value{
The adjusted TMB value and correlation figure
}
\description{
Correct panel TMB evaluation by modeling adjustment.
}
\examples{
TMBpredict("COAD","sample.vcf","panel.bed","panel2wes")
}
